Fixed the package names and renamed book for the sake of consistency.
Title
