+++ bookstore/src/main/java/app/bookstore/author.java
@@ -1,4 +1,4 @@
-package geekText;
+package app.bookstore;
 
 public class author 
 {
